+++ src/com/onsoft/glasscat/jslets/utils/JsletContextBuilder.ts
@@ -16,7 +16,7 @@
 
 import {Jslet, JsletContext, SecurityContext, SessionContext} from "jec-exchange";
 import {EjpJsletContext} from "../EjpJsletContext";
-import {GlobalClassLoader} from "jec-commons";
+import {GlobalClassLoader, JecStringsEnum, UrlStringsEnum, PathUtils} from "jec-commons";
 import {LocaleManager} from "jec-commons-node";
 import {GlassCatLocaleManager} from "../../i18n/GlassCatLocaleManager";
 import {LoggerManager} from "../../util/logging/LoggerManager";
@@ -24,6 +24,8 @@ import {DomainConnector} from "../../domains/connectors/DomainConnector";
 import {LoginStrategy} from "../../security/login/LoginStrategy";
 import {GlassCatError} from "../../exceptions/GlassCatError";
 import {GlassCatErrorCode} from "../../exceptions/GlassCatErrorCode";
+import {SokokeMetadataRefs, Sokoke} from "jec-sokoke";
+import {InjectionPoint, Bean, ScopeType, Scope} from "jec-jdi";
 
 /**
  * The singleton helper for creating <code>JsletContext</code> objects, from 
@@ -80,6 +82,41 @@ export class JsletContextBuilder {
   //////////////////////////////////////////////////////////////////////////////
 
   /**
+   * HACK! Performs dependency injection for the jslet fields.
+   * 
+   * @param {any} jslet the jslet for which to perform dependency injection.
+   */
+  private performDI(jslet:any):void {
+    const sokoke:Sokoke = (Sokoke.getInstance() as Sokoke);
+    const injectionPoints:Array<InjectionPoint> =
+                      jslet[SokokeMetadataRefs.SOKOKE_INJECTION_POINT_METADATA];
+    let len:number = -1;
+    let injectionPoint:InjectionPoint = null;
+    let value:any = null;
+    let bean:Bean = null;
+    let scopeType:ScopeType = null;
+    let scope:Scope = null;
+    if(injectionPoints) {
+      len = injectionPoints.length;
+      while(len--) {
+        injectionPoint = injectionPoints[len];
+        bean = injectionPoint.getBean();
+        if(bean) {
+          scope = bean.getScope();
+          scopeType = scope ? scope.getType() : ScopeType.DEPENDENT;
+          if(scopeType === ScopeType.APPLICATION ||
+             scopeType === ScopeType.DEPENDENT) {
+              value = sokoke.getInjectableReference(bean);
+              Reflect.defineProperty(
+                jslet, injectionPoint.getElement().getName(), { value: value }
+              );
+          }
+        }
+      }
+    }
+  }
+
+  /**
    * Creates and returns a new <code>Jslet</code> instance.
    *
    * @param {string} path the path to the jslet class file.
@@ -87,14 +124,16 @@ export class JsletContextBuilder {
    * @return {Jslet} a new <code>Jslet</code> instance.
    */
   private buildJslet(path:string, target:string):Jslet {
+    const filePath:string = PathUtils.getInstance().buildFilePath(target, path);
     let jslet:Jslet = null;
     let Contructor:any = null;
     try {
-      Contructor = GlobalClassLoader.getInstance().loadClass(target + path);
+      Contructor = GlobalClassLoader.getInstance().loadClass(filePath);
       jslet = new Contructor();
+      this.performDI(jslet);
     } catch(e){
-      // TODO: log error message
-      throw new GlassCatError(GlassCatErrorCode.INVALID_JSLET_CONFIG);
+      // TODO: correctly log error message
+      throw new GlassCatError(GlassCatErrorCode.INVALID_JSLET_CONFIG, e);
     }
     return jslet;
   }