+++ utils/test-utils/utilities/BasicSessionErrorTestUtils.ts
@@ -14,7 +14,6 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-import { LogLevelUtil } from "jec-commons";
 import { GlassCatSessionId } from "../../../src/com/onsoft/glasscat/security/session/GlassCatSessionId";
 import { SessionId, SessionErrorType } from "jec-exchange";